+++ test/mocks.dart
@@ -3,5 +3,9 @@ import 'package:cafe_bda/services/google_sheets_service.dart';
 import 'package:cafe_bda/repositories/cafe_repository.dart';
 import 'package:cafe_bda/providers/cafe_data_provider.dart';
 
-@GenerateMocks([GoogleSheetsService, CafeRepository, CafeDataProvider])
+@GenerateNiceMocks([
+  MockSpec<GoogleSheetsService>(),
+  MockSpec<CafeRepository>(),
+  MockSpec<CafeDataProvider>(),
+])
 void main() {}